Build a class for formatting data into strings
Make a stream-like formatting class which allows you to store references
to variables and then retrieve the format later as a string. Will
hopefully be used for adding dynamic text on a HUD by storing the
variables it represents. I need a way to do things like setting width
and stuff, but it works for now.