Add a more general TextEntity class
Now, FormattedTextEntity inherits from TextEntity. There is the minor
annoyance that TextEntity has to store the text that it is about to
draw, so you need to reset the stored text every frame so that it draws
correctly. I guess this isn't that big of a deal, because it was being
changed anyway, but it seems a little sloppy.